refactor: minimise duplicates of test KeyringSigner generators
