+++ app/estimate_square_size_test.go
@@ -36,7 +36,7 @@ func Test_estimateSquareSize(t *testing.T) {
 		{"one over the perfect estimation edge case", 10, 1, appconsts.SparseShareContentSize * 10, 8},
 	}
 	encConf := encoding.MakeConfig(ModuleEncodingRegisters...)
-	signer := generateKeyringSigner(t, "estimate-key")
+	signer := GenerateKeyringSigner(t, "estimate-key")
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			txs := generateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
@@ -70,7 +70,7 @@ func Test_estimateSquareSize(t *testing.T) {
 
 func Test_pruning(t *testing.T) {
 	encConf := encoding.MakeConfig(ModuleEncodingRegisters...)
-	signer := generateKeyringSigner(t, "estimate-key")
+	signer := GenerateKeyringSigner(t, "estimate-key")
 	txs := generateManyRawSendTxs(t, encConf.TxConfig, signer, 10)
 	txs = append(txs, generateManyRawWirePFB(t, encConf.TxConfig, signer, 10, 1000)...)
 	parsedTxs := parseTxs(encConf.TxConfig, txs)
@@ -125,7 +125,7 @@ func Test_overEstimateMalleatedTxSize(t *testing.T) {
 	}
 
 	encConf := encoding.MakeConfig(ModuleEncodingRegisters...)
-	signer := generateKeyringSigner(t, "estimate-key")
+	signer := GenerateKeyringSigner(t, "estimate-key")
 	for _, tt := range tests {
 		wpfbTx := generateRawWirePFBTx(
 			t,
@@ -163,7 +163,7 @@ func Test_calculateCompactShareCount(t *testing.T) {
 		{"one over the perfect estimation edge case", 10, 1, totalBlobSize(appconsts.SparseShareContentSize + 1)},
 	}
 	encConf := encoding.MakeConfig(ModuleEncodingRegisters...)
-	signer := generateKeyringSigner(t, "estimate-key")
+	signer := GenerateKeyringSigner(t, "estimate-key")
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			txs := generateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
