+++ x/blob/types/builder_test.go
@@ -5,8 +5,6 @@ import (
 	"testing"
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
-	"github.com/cosmos/cosmos-sdk/crypto/hd"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/tx"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
@@ -22,7 +20,7 @@ func TestBuildWirePayForBlob(t *testing.T) {
 	info, err := testRing.Key(testAccName)
 	require.NoError(t, err)
 
-	k := NewKeyringSigner(testRing, testAccName, "chain-id")
+	k := NewKeyringSigner(testRing, testAccName, testChainID)
 	require.NoError(t, err)
 
 	namespace := []byte{1, 1, 1, 1, 1, 1, 1, 1}
@@ -64,7 +62,7 @@ func TestBroadcastPayForBlob(t *testing.T) {
 	require.NoError(t, err)
 	t.Skipf("no local connection to app and no funds in wallet %s", addr)
 
-	k := NewKeyringSigner(testRing, testAccName, "test")
+	k := NewKeyringSigner(testRing, testAccName, testChainID)
 
 	RPCAddress := "127.0.0.1:9090"
 
@@ -103,9 +101,7 @@ func TestBroadcastPayForBlob(t *testing.T) {
 
 func TestQueryAccountNumber(t *testing.T) {
 	t.Skip("no local connection to app and no funds in wallet")
-	testRing := generateKeyring(t)
-
-	k := NewKeyringSigner(testRing, testAccName, "test")
+	k := generateKeyringSigner(t, testAccName)
 
 	RPCAddress := "127.0.0.1:9090"
 
@@ -114,35 +110,3 @@ func TestQueryAccountNumber(t *testing.T) {
 	err = k.QueryAccountNumber(context.TODO(), rpcClient)
 	require.NoError(t, err)
 }
-
-func generateKeyring(t *testing.T, accts ...string) keyring.Keyring {
-	t.Helper()
-	encCfg := makeBlobEncodingConfig()
-	kb := keyring.NewInMemory(encCfg.Codec)
-
-	for _, acc := range accts {
-		_, _, err := kb.NewMnemonic(acc, keyring.English, "", "", hd.Secp256k1)
-		if err != nil {
-			t.Error(err)
-		}
-	}
-
-	_, err := kb.NewAccount(testAccName, testMnemo, "1234", "", hd.Secp256k1)
-	if err != nil {
-		panic(err)
-	}
-
-	return kb
-}
-
-func generateKeyringSigner(t *testing.T, accts ...string) *KeyringSigner {
-	kr := generateKeyring(t, accts...)
-	return NewKeyringSigner(kr, testAccName, testChainID)
-}
-
-const (
-	// nolint:lll
-	testMnemo   = `ramp soldier connect gadget domain mutual staff unusual first midnight iron good deputy wage vehicle mutual spike unlock rocket delay hundred script tumble choose`
-	testAccName = "test-account"
-	testChainID = "test-chain-1"
-)
